chore: environment configurations

+++ build/clean.js
@@ -1,9 +1,12 @@
 var fs = require('fs');
 var glob = require('glob');
 
-glob('dist/**/*.*', function(err, matches) {
-    matches.forEach(function(filepath) {
-        fs.unlinkSync(filepath);
+if(fs.existsSync('dist/')) {
+    glob('dist/**/*.*', function(err, matches) {
+        matches.forEach(function(filepath) {
+            fs.unlinkSync(filepath);
+        });
+        fs.rmdirSync('dist/');
     });
-    fs.rmdirSync('dist/');
-});
+}
+
